chore: add voter adapter + fix tests
