+++ decidim-bulletin_board-app/app/services/voting_scheme.rb
@@ -3,6 +3,9 @@
 require "voting_scheme/bulletin_board"
 require "voting_scheme/dummy/bulletin_board"
 require "voting_scheme/electionguard/bulletin_board"
+require "voting_scheme/voter"
+require "voting_scheme/dummy/voter"
+require "voting_scheme/electionguard/voter"
 
 module VotingScheme
   VOTING_SCHEME = {
@@ -12,11 +15,16 @@ module VotingScheme
 
   def self.from_name(name)
     {
-      bulletin_board: "#{VOTING_SCHEME[name.to_sym]}::BulletinBoard".constantize
+      bulletin_board: bulletin_board_adapter(name),
+      voter: voter_adapter(name)
     }
   end
 
-  def self.results_message?(voting_scheme_name, type_subtype)
-    from_name(voting_scheme_name)::RESULTS.include?(type_subtype)
+  def self.bulletin_board_adapter(scheme_name)
+    "#{VOTING_SCHEME[scheme_name.to_sym]}::BulletinBoard".constantize if VOTING_SCHEME[scheme_name.to_sym].present?
+  end
+
+  def self.voter_adapter(scheme_name)
+    "#{VOTING_SCHEME[scheme_name.to_sym]}::Voter".constantize if VOTING_SCHEME[scheme_name.to_sym].present?
   end
 end
