+++ decidim-bulletin_board-app/app/services/voting_scheme/voter.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module VotingScheme
+  class Voter
+    def initialize(election)
+      @election = election
+    end
+
+    attr_reader :election
+  end
+
+  class RejectedMessage < StandardError
+  end
+end
